+++ src/features/auth/routes/login.tsx
@@ -1,6 +1,6 @@
 
 import { useNavigate } from 'react-router-dom';
-import { LoginForm } from '../components/__tests__/login-form';
+import { LoginForm } from '../components/login-form';
 import { useUser } from 'hooks/use-user';
 import { useEffect } from 'react';
